+++ deltafi-common/src/test/java/org/deltafi/common/cache/TestService.java
@@ -1,0 +1,41 @@
+/*
+ *    DeltaFi - Data transformation and enrichment platform
+ *
+ *    Copyright 2021-2023 DeltaFi Contributors <deltafi@deltafi.org>
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+package org.deltafi.common.cache;
+
+import lombok.Data;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.stereotype.Component;
+
+@Component
+@Data
+public class TestService {
+    private int lookupACount;
+    private int lookupBCount;
+
+    @Cacheable("test-cache-a")
+    public String lookupA(String key) {
+        lookupACount++;
+        return key + "-a";
+    }
+
+    @Cacheable("test-cache-b")
+    public String lookupB(String key) {
+        lookupBCount++;
+        return key + "-b";
+    }
+}